Changed the color scheme of the alerts to increase contrast
This is based on contrast ratio, staying close to the recommendations for visual-audio contrast recommendations from W3C (minimum 4.5:1 for small text, 7:1 for small text and users with a lower visual acuity).
Red alerts stayed the same bright red for now, as the emphasis is for the alert to be immedialty recognized as important, but changed the text color from white to black for better contrast.